Crate datamodel

This crate is responsible for parsing, rendering and formatting a Prisma Schema. A Prisma Schema consists out of two parts:

The Datamodel part refers to the model and enum definitions.

The Configuration part refers to the generator and datasource definitions.

The data structures are organized into 3 layers:

The AST layer contains the data data structures representing the schema input.

The model layer contains the data structures that are semantically rich and therefore engines can build upon.

The JSON layer contains the data structures that represent the contract for the DMMF which is the API for client generators.

The responsibilities of each top level module is as following:

common: contains constants and generic helpers

error: contains the error and result types

ast: contains the data structures for the AST of a Prisma schema. And the parsing functions to turn an input string into an AST.

dml: contains the models representing the Datamodel part of a Prisma schema

configuration: contains the models representing the Datasources and Generators of a Prisma schema

transform: contains the logic to turn an AST into models and vice versa

mcf: contains the logic to turn generators and datasources into their JSON representation

The flow between the layers is depicted in the following diagram.

                ┌──────────────────┐
                │       json       │
                └────────▲─────────┘
                         │
                         │
      ┌──────────────────┐┌──────────────────┐
      │       dml        ││  configuration   │
      └──────────────────┘└──────────────────┘
                        │  ▲
┌─────────────────────┐ │  │ ┌─────────────────────┐
│transform::dml_to_ast│ │  │ │transform::ast_to_dml│
└─────────────────────┘ │  │ └─────────────────────┘
                        │  │
                        ▼  │
                ┌──────────────────┐
                │       ast        │
                └──────────────────┘
                        │  ▲
                        │  │
                        │  │
                        ▼  │
                 ┌──────────────────┐
                 │  schema string   │
                 └──────────────────┘

Re-exports

pub use datamodel_connector;

pub use diagnostics;

pub use parser_database;

pub use schema_ast;

Modules

ast

The AST data structure. It aims to faithfully represent the syntax of a Prisma Schema, with source span information.

common

This module contains shared constants and logic that can be used by engines.

dml

mcf

mcf: Turns a collection of configuration::Datasource and configuration::Generator into a JSON representation.

Structs

Configuration

Datasource

a datasource from the prisma schema.

Generator

StringFromEnvVar

Either an env var or a string literal.

Validated

ValidatedSchema

Functions

is_reserved_type_name

Is this a valid type name for the Prisma client API?

parse_configuration

Loads all configuration blocks from a datamodel using the built-in source definitions.

parse_datamodel

Parses and validates a datamodel string, using core attributes only.

parse_schema

Parse and validate the whole schema

parse_schema_ast

parse_schema_parserdb

Parse and validate the whole schema.

reformat

Returns either the reformatted schema, or the original input if we can’t reformat. This happens if and only if the source does not parse to a valid AST.

render_datamodel_and_config_to_string

Renders a datamodel, sources and generators.

render_datamodel_to_string

Renders the datamodel without configuration blocks.

Type Definitions

ValidatedConfiguration

ValidatedDatamodel