Module datamodel::walkers

This module contains convenience functions to easily traverse a Datamodel tree. The most prominent functionality is the pain free navigation of relations.

Structs

EnumWalker

IndexWalker

ManyToManyRelationWalker

ModelWalker

RelationFieldWalker

RelationWalker

ScalarFieldWalker

Enums

TypeWalker

Functions

find_model_by_db_name

Find the model mapping to the passed in database name.

walk_models

Iterator over all the models in the schema.

walk_relations

Iterator over all the relations in the schema. Each relation will only occur once.

walk_scalar_fields

Iterator to walk all the scalar fields in the schema, associating them with their parent model.