Crate datamodel

PSL — Prisma Schema Language

This crate is responsible for parsing, validating, formatting and rendering a PSL documents.

A Prisma Schema consists out of two parts:

The Datamodel part refers to the model and enum definitions.

The Configuration part refers to the generator and datasource definitions.

Re-exports

pub use datamodel_connector;

pub use diagnostics;

pub use dml;

pub use parser_database;

Modules

builtin_connectors

common

This module contains shared constants and logic that can be used by engines.

mcf

mcf: Turns a collection of configuration::Datasource and configuration::Generator into a JSON representation.

Structs

Configuration

Datasource

a datasource from the prisma schema.

Generator

StringFromEnvVar

Either an env var or a string literal.

Validated

ValidatedSchema

Functions

is_reserved_type_name

Is this a valid type name for the Prisma client API?

parse_configuration

Loads all configuration blocks from a datamodel using the built-in source definitions.

parse_datamodel

Parses and validates a datamodel string, using core attributes only.

parse_schema

Parse and validate the whole schema

parse_schema_ast

parse_schema_parserdb

reformat

Returns either the reformatted schema, or the original input if we can’t reformat. This happens if and only if the source does not parse to a well formed AST.

render_datamodel_and_config_to_string

Renders a datamodel, sources and generators.

render_datamodel_to_string

Renders the datamodel without configuration blocks.

Type Definitions

ValidatedConfiguration

ValidatedDatamodel