Crate datamodel

Prisma Schema Language implementation

Overview

This crate is the public API for working with Prisma schemas: parsing and validating their string representation, analyzing them, providing convenient APIs, rendering, reformatting, etc. It is the main implementation of the Prisma Schema Language, and it is used by all Prisma engines in this repository.

The query engine further processes Prisma schemas to produce the client API, and the DMMF JSON format.

Design goals

Strict parsing: a duplicate attribute, unknown attribute, unknown argument or extra argument is an error.

Expose a convenient and obvious / hard-to-misuse public API.

Expose the source information (AST node spans, etc) in the parsed schema.

Accumulate errors to present them at the end instead of returning early.

Usage

Please see lib.rs and the rustdoc documentation for the public API.

Main use-case, parsing a string to datamodel:

let file = std::fs::read_to_string(&args[1]).unwrap();
let validated_schema = datamodel::parse_schema_parserdb(&file)?;

Re-exports

pub use datamodel_connector;

pub use diagnostics;

pub use dml;

pub use parser_database;

Modules

builtin_connectors

common

This module contains shared constants and logic that can be used by engines.

mcf

mcf: Turns a collection of configuration::Datasource and configuration::Generator into a JSON representation. This is the get_config() representation.

Structs

Configuration

Datasource

a datasource from the prisma schema.

Generator

StringFromEnvVar

Either an env var or a string literal.

Validated

ValidatedSchema

Functions

is_reserved_type_name

Is this a valid type name for the Prisma client API?

parse_configuration

Loads all configuration blocks from a datamodel using the built-in source definitions.

parse_datamodel

Parses and validates a datamodel string, using core attributes only.

parse_schema

Parse and validate the whole schema

parse_schema_ast

parse_schema_parserdb

reformat

Returns either the reformatted schema, or the original input if we can’t reformat. This happens if and only if the source does not parse to a well formed AST.

render_datamodel_and_config_to_string

Renders a datamodel, sources and generators.

render_datamodel_to_string

Renders the datamodel without configuration blocks.

Type Definitions

ValidatedConfiguration

ValidatedDatamodel